backend/requirements.txt
feedparser==6.0.10
yfinance==0.2.32
google-generativeai==0.3.0
requests==2.31.0
python-dotenv==1.0.0
